% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactCodeFit.R
\name{impactCodeFit}
\alias{impactCodeFit}
\title{impactCodeFit}
\usage{
impactCodeFit(mydata, catCols, classCol, fam = "guassian")
}
\arguments{
\item{mydata}{a data frame}

\item{catCols}{names of a grouped catagories in the mydata}

\item{classCol}{outcome name use to create impact codes}

\item{family}{a glm family}
}
\value{
a list of intercept only multi level models to be used to transform new data
}
\description{
creates a multi level intercept model for a single catagory using glmer from lmer4
}
