% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactCodeTransform.R
\name{impactCodeTransform}
\alias{impactCodeTransform}
\title{impactCodeTransform}
\usage{
impactCodeTransform(impactCodes, newData)
}
\arguments{
\item{newData}{a data frame}

\item{impactCodeFit}{a model list generated from impactCodeFit}
}
\value{
a data frame with the catagory levels replaced with credibility wieghted averages
}
\description{
applies a list from impactCodeFit transformations to unseen data
}
