% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleVarImpactCode.R
\name{singleVarImpactCode}
\alias{singleVarImpactCode}
\title{singleVarImpactCode}
\usage{
singleVarImpactCode(mydata, catCol = "MSDRGName",
  classCol = "UHC_ReadmissionFlag", fam = "gaussian")
}
\arguments{
\item{mydata}{a data frame}

\item{catCol}{name of a grouped catagory in the mydata}

\item{classCol}{outcome name use to create impact codes}
}
\description{
creates a multi level intercept model for a single catagory using glmer from lmer4
}
